refactor: change Integer to primitive types in response DTOs and repositories

+++ src/main/java/com/fitassist/backend/dto/request/activity/CalculateActivityCaloriesRequestDto.java
@@ -17,7 +17,7 @@ public class CalculateActivityCaloriesRequestDto {
 
 	@NotNull
 	@Positive
-	private int time;
+	private Short time;
 
 	@Positive
 	private BigDecimal weight;
